+++ src/main/scala/github/gphat/censorinus/DogStatsDClient.scala
@@ -37,62 +37,117 @@ class DogStatsDClient(
     * @param name The name of the metric
     * @param value The value of the metric, or how much to increment by
     * @param sampleRate The rate at which to sample this metric.
+    * @param bypassSampler If true, the metric will always be passed through, but the sample rate will be included in the emitted metric. This is useful for when you occasionally do your own sampling.
     */
-  def counter(name: String, value: Double, sampleRate: Double = defaultSampleRate, tags: Seq[String] = Seq.empty) = enqueue(
-    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "c", tags = tags)
+  def counter(
+    name: String,
+    value: Double,
+    sampleRate: Double = defaultSampleRate,
+    tags: Seq[String] = Seq.empty,
+    bypassSampler: Boolean = false
+  ) = enqueue(
+    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "c", tags = tags),
+    sampleRate,
+    bypassSampler
   )
 
   /** Emit a decrement metric.
     * @param name The name of the metric
     * @param value The value of the metric, or how much to decrement by. Defaults to -1
     * @param sampleRate The rate at which to sample this metric.
+    * @param bypassSampler If true, the metric will always be passed through, but the sample rate will be included in the emitted metric. This is useful for when you occasionally do your own sampling.
     */
-  def decrement(name: String, value: Double = 1, sampleRate: Double = defaultSampleRate, tags: Seq[String] = Seq.empty) = enqueue(
-    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "c", tags = tags)
+  def decrement(
+    name: String,
+    value: Double = 1,
+    sampleRate: Double = defaultSampleRate,
+    tags: Seq[String] = Seq.empty,
+    bypassSampler: Boolean = false
+  ) = enqueue(
+    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "c", tags = tags),
+    sampleRate,
+    bypassSampler
   )
 
   /** Emit a gauge metric.
     * @param name The name of the metric
     * @param value The value of the metric, or current value of the gauge
     * @param sampleRate The rate at which to sample this metric.
+    * @param bypassSampler If true, the metric will always be passed through, but the sample rate will be included in the emitted metric. This is useful for when you occasionally do your own sampling.
     */
-  def gauge(name: String, value: Double, sampleRate: Double = defaultSampleRate, tags: Seq[String] = Seq.empty) = enqueue(
-    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "g", tags = tags)
+  def gauge(
+    name: String,
+    value: Double,
+    sampleRate: Double = defaultSampleRate,
+    tags: Seq[String] = Seq.empty,
+    bypassSampler: Boolean = false
+  ) = enqueue(
+    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "g", tags = tags),
+    sampleRate,
+    bypassSampler
   )
 
   /** Emit a histogram metric.
     * @param name The name of the metric
     * @param value The value of the metric, or a value to be sampled for the histogram
     * @param sampleRate The rate at which to sample this metric.
+    * @param bypassSampler If true, the metric will always be passed through, but the sample rate will be included in the emitted metric. This is useful for when you occasionally do your own sampling.
     */
-  def histogram(name: String, value: Double, sampleRate: Double = defaultSampleRate, tags: Seq[String] = Seq.empty) = enqueue(
-    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "h", tags = tags)
+  def histogram(
+    name: String,
+    value: Double,
+    sampleRate: Double = defaultSampleRate,
+    tags: Seq[String] = Seq.empty,
+    bypassSampler: Boolean = false) = enqueue(
+    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "h", tags = tags),
+    sampleRate,
+    bypassSampler
   )
 
   /** Emit an increment metric.
     * @param name The name of the metric
     * @param value The value of the metric, or the amount to increment by. Defaults to 1
     * @param sampleRate The rate at which to sample this metric.
+    * @param bypassSampler If true, the metric will always be passed through, but the sample rate will be included in the emitted metric. This is useful for when you occasionally do your own sampling.
     */
-  def increment(name: String, value: Double = 1, sampleRate: Double = defaultSampleRate, tags: Seq[String] = Seq.empty) = enqueue(
-    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "c", tags = tags)
+  def increment(
+    name: String, value: Double = 1,
+    sampleRate: Double = defaultSampleRate,
+    tags: Seq[String] = Seq.empty,
+    bypassSampler: Boolean = false
+  ) = enqueue(
+    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "c", tags = tags),
+    sampleRate,
+    bypassSampler
   )
 
   /** Emit a meter metric.
     * @param name The name of the metric
     * @param value The value of the meter
     * @param sampleRate The rate at which to sample this metric.
+    * @param bypassSampler If true, the metric will always be passed through, but the sample rate will be included in the emitted metric. This is useful for when you occasionally do your own sampling.
     */
-  def meter(name: String, value: Double, sampleRate: Double = defaultSampleRate, tags: Seq[String] = Seq.empty) = enqueue(
-    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "m", tags = tags)
+  def meter(
+    name: String,
+    value: Double,
+    sampleRate: Double = defaultSampleRate,
+    tags: Seq[String] = Seq.empty,
+    bypassSampler: Boolean = false
+  ) = enqueue(
+    Metric(name = makeName(name), value = floatFormat.format(value), sampleRate = sampleRate, metricType = "m", tags = tags),
+    sampleRate,
+    bypassSampler
   )
 
   /** Emit a set metric.
    * @param name The name of the metric
    * @param value The item to add to the set
-   * @param sampleRate The rate at which to sample the metric
    */
-  def set(name: String, value: String, tags: Seq[String] = Seq.empty) = enqueue(
+  def set(
+    name: String,
+    value: String,
+    tags: Seq[String] = Seq.empty
+  ) = enqueue(
     Metric(name = makeName(name), value = value, metricType = "s", tags = tags)
   )
 
@@ -100,8 +155,17 @@ class DogStatsDClient(
     * @param name The name of the metric
     * @param value The value of the timer in milliseconds
     * @param sampleRate The rate at which to sample this metric.
+    * @param bypassSampler If true, the metric will always be passed through, but the sample rate will be included in the emitted metric. This is useful for when you occasionally do your own sampling.
     */
-  def timer(name: String, milliseconds: Double, sampleRate: Double = defaultSampleRate, tags: Seq[String] = Seq.empty) = enqueue(
-    Metric(name = makeName(name), value = floatFormat.format(milliseconds), sampleRate = sampleRate, metricType = "ms", tags = tags)
+  def timer(
+    name: String,
+    milliseconds: Double,
+    sampleRate: Double = defaultSampleRate,
+    tags: Seq[String] = Seq.empty,
+    bypassSampler: Boolean = false
+  ) = enqueue(
+    Metric(name = makeName(name), value = floatFormat.format(milliseconds), sampleRate = sampleRate, metricType = "ms", tags = tags),
+    sampleRate,
+    bypassSampler
   )
 }